+++ zone/raids.cpp
@@ -470,7 +470,7 @@ void Raid::HealGroup(uint32 heal_amt, Mob* caster, uint32 gid, int32 range)
 		range = 200;
 
 	float distance;
-	float range2 = range*range;
+	float range2 = (float)(range*range);
 
 	int numMem = 0;
 	unsigned int gi = 0;
@@ -516,7 +516,7 @@ void Raid::BalanceHP(int32 penalty, uint32 gid, int32 range, Mob* caster)
 	int gi = 0;
 	
 	float distance;
-	float range2 = range*range;
+	float range2 = (float)(range*range);
 
 	for(; gi < MAX_RAID_MEMBERS; gi++)
 	{
@@ -564,7 +564,7 @@ void Raid::BalanceMana(int32 penalty, uint32 gid, int32 range, Mob* caster)
 		range = 200;
 			
 	float distance;
-	float range2 = range*range;
+	float range2 = (float)(range*range);
 
 	int manataken = 0, numMem = 0;
 	int gi = 0;
