+++ zone/zone.h
@@ -243,7 +243,7 @@ public:
 	uint8 lootvar;
 
 	bool	HasGraveyard();
-	void	SetGraveyard(uint32 zoneid, uint32 x, uint32 y, uint32 z, uint32 heading);
+	void	SetGraveyard(uint32 zoneid, float x, float y, float z, float heading);
 
 	void		LoadBlockedSpells(uint32 zoneid);
 	void		ClearBlockedSpells();
